+++ examples/multiple-renderer/renderer/foo.ts
@@ -1,0 +1,7 @@
+import { routes } from './router'
+
+document.getElementById('app')!.innerHTML = `
+<h1>examples/multiple-renderer/foo.ts</h1>
+<hr>
+${routes()}
+`