+++ examples/multiple-window/electron/preload.ts
@@ -1,0 +1,10 @@
+import { contextBridge, ipcRenderer } from 'electron'
+
+contextBridge.exposeInMainWorld(
+  'electron',
+  {
+    toWindow(html: string) {
+      ipcRenderer.invoke('to-window', html)
+    },
+  },
+)