+++ examples/multiple-window/renderer/index.ts
@@ -1,0 +1,4 @@
+document.getElementById('app')!.innerHTML = `
+<h1>examples/multiple-window/index.ts</h1>
+<hr>
+`