+++ examples/multiple-window/vite.node-false.config.ts
@@ -1,0 +1,17 @@
+import path from 'node:path'
+import { defineConfig } from 'vite'
+import renderer from 'vite-plugin-electron-renderer'
+
+export default defineConfig({
+  plugins: [
+    renderer(),
+  ],
+  build: {
+    outDir: 'dist',
+    emptyOutDir: false,
+    minify: false,
+    rollupOptions: {
+      input: path.join(__dirname, 'html/node-false.html'),
+    },
+  },
+})